renamed contact listener file and removed end contact function as its not really needed with the current implementation.
